+++ packages/office-ui-fabric-react/src/components/SpinButton/index.ts
@@ -1,2 +1,3 @@
 export * from './SpinButton';
+export * from './SpinButton.base';
 export * from './SpinButton.types';
